+++ src/main/java/com/waimai/dao/ResourceDao.java
@@ -1,6 +1,7 @@
 package com.waimai.dao;
 
 import java.util.List;
+
 import com.waimai.model.Menu;
 import com.waimai.model.Resource;
 
@@ -36,7 +37,7 @@ public interface ResourceDao {
 	 * @Author: 李年
 	 * @CreateDate: 2013-3-28
 	 */
-	public List<Resource> findResourceByParentId(Long menuId);
+	public List<Resource> findResourceByParentId(Integer menuId);
 	/**
 	 * @FunName: findResourceByParentId
 	 * @Description:  通过二级菜单获取全部三级资源
@@ -45,7 +46,7 @@ public interface ResourceDao {
 	 * @Author: 李年
 	 * @CreateDate: 2013-3-28
 	 */
-	public List<Resource> findAllResourceByParentId(Long menuId);
+	public List<Resource> findAllResourceByParentId(Integer menuId);
 	/**
 	 * @FunName: findResourceByRole
 	 * @Description:  通过角色拿到所有资源
@@ -64,5 +65,5 @@ public interface ResourceDao {
 	 * @Author: 李年
 	 * @CreateDate: 2013-3-28
 	 */
-	public Resource loadResourceByResourceId(Long id,Menu menu);
+	public Resource loadResourceByResourceId(Integer id,Menu menu);
 }
